refactor: Initialize the NumberGuess class without the guess argument.
Initializing the class each time with the guess of the user forces
new classes to be made per guess. This isn't very helpful in tracking
if the guess is too high or too low so initializing the class with an
optional "high" value made more sense. Now the user can decide to what
the highest number that the computer can generate.